fixed my careless ctrl-f ing

+++ assignment1/flex/lexer.h
@@ -90,7 +90,7 @@ enum tokens {
 	CONTINUE,
 	DEFAULT,
 	DO,
-	longlong,
+	DOUBLE,
 	ELSE,
 	ENUM,
 	EXTERN,
